requirements.txt
Flask>=3.0,<4.0
flask-sock>=0.7,<1.0
python-dotenv>=1.0,<2.0
pyserial>=3.5,<4.0
